requirements.txt
numpy~=2.2.5
torch~=2.7.0
tqdm~=4.67.1
pandas~=2.2.3
pillow~=11.2.1
setuptools~=79.0.1
ultralytics-thop~=2.0.14
requests~=2.32.3
torchvision~=0.22.0
psutil~=7.0.0
matplotlib~=3.10.1
pyyaml~=6.0.2
scipy~=1.15.2
scikit-learn~=1.6.1